+++ src/com/maddyhome/idea/vim/group/SearchGroup.java
@@ -22,6 +22,7 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.*;
 import com.intellij.openapi.editor.colors.EditorColors;
+import com.intellij.openapi.editor.colors.EditorColorsScheme;
 import com.intellij.openapi.editor.event.DocumentEvent;
 import com.intellij.openapi.editor.event.DocumentListener;
 import com.intellij.openapi.editor.markup.*;
@@ -30,6 +31,8 @@ import com.intellij.openapi.fileEditor.FileEditorManagerEvent;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManager;
 import com.intellij.openapi.util.Ref;
+import com.intellij.ui.ColorUtil;
+import com.intellij.util.Processor;
 import com.maddyhome.idea.vim.VimPlugin;
 import com.maddyhome.idea.vim.command.CommandFlags;
 import com.maddyhome.idea.vim.command.SelectionType;
@@ -50,6 +53,7 @@ import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import javax.swing.*;
 import java.awt.*;
 import java.text.NumberFormat;
 import java.text.ParsePosition;
@@ -163,34 +167,40 @@ public class SearchGroup {
   @NotNull
   private static ReplaceConfirmationChoice confirmChoice(@NotNull Editor editor, @NotNull String match) {
     final Ref<ReplaceConfirmationChoice> result = Ref.create(ReplaceConfirmationChoice.QUIT);
-    // XXX: The Ex entry panel is used only for UI here, its logic might be inappropriate for this method
-    final ExEntryPanel exEntryPanel = ExEntryPanel.getInstanceWithoutShortcuts();
-    exEntryPanel.activate(editor, new EditorDataContext(editor), "Replace with " + match + " (y/n/a/q/l)?", "", 1);
-    ModalEntry.activate(key -> {
+    final Processor<KeyStroke> keyStrokeProcessor = key -> {
       final ReplaceConfirmationChoice choice;
       final char c = key.getKeyChar();
       if (StringHelper.isCloseKeyStroke(key) || c == 'q') {
         choice = ReplaceConfirmationChoice.QUIT;
-      }
-      else if (c == 'y') {
+      } else if (c == 'y') {
         choice = ReplaceConfirmationChoice.SUBSTITUTE_THIS;
-      }
-      else if (c == 'l') {
+      } else if (c == 'l') {
         choice = ReplaceConfirmationChoice.SUBSTITUTE_LAST;
-      }
-      else if (c == 'n') {
+      } else if (c == 'n') {
         choice = ReplaceConfirmationChoice.SKIP;
-      }
-      else if (c == 'a') {
+      } else if (c == 'a') {
         choice = ReplaceConfirmationChoice.SUBSTITUTE_ALL;
-      }
-      else {
+      } else {
         return true;
       }
       result.set(choice);
-      exEntryPanel.deactivate(true);
       return false;
-    });
+    };
+    if (ApplicationManager.getApplication().isUnitTestMode()) {
+      final TestInputModel inputModel = TestInputModel.getInstance(editor);
+      for (KeyStroke key = inputModel.nextKeyStroke(); key != null; key = inputModel.nextKeyStroke()) {
+        if (!keyStrokeProcessor.process(key)) {
+          break;
+        }
+      }
+    }
+    else {
+      // XXX: The Ex entry panel is used only for UI here, its logic might be inappropriate for this method
+      final ExEntryPanel exEntryPanel = ExEntryPanel.getInstanceWithoutShortcuts();
+      exEntryPanel.activate(editor, new EditorDataContext(editor), "Replace with " + match + " (y/n/a/q/l)?", "", 1);
+      ModalEntry.activate(keyStrokeProcessor);
+      exEntryPanel.deactivate(true, false);
+    }
     return result.get();
   }
 
@@ -324,7 +334,15 @@ public class SearchGroup {
   private int searchNextWithDirection(@NotNull Editor editor, @NotNull Caret caret, int count, int dir) {
     resetShowSearchHighlight();
     updateSearchHighlights();
-    return findItOffset(editor, caret.getOffset(), count, dir);
+    final int startOffset = caret.getOffset();
+    int offset = findItOffset(editor, startOffset, count, dir);
+    if (offset == startOffset) {
+      /* Avoid getting stuck on the current cursor position, which can
+       * happen when an offset is given and the cursor is on the last char
+       * in the buffer: Repeat with count + 1. */
+      offset = findItOffset(editor, startOffset, count + 1, dir);
+    }
+    return offset;
   }
 
   private void resetShowSearchHighlight() {
@@ -348,14 +366,10 @@ public class SearchGroup {
     updateSearchHighlights(lastSearch, lastIgnoreSmartCase, showSearchHighlight, true);
   }
 
-  public void updateIncsearchHighlights(@NotNull Editor editor, @NotNull String pattern, boolean forwards, int caretOffset, @Nullable LineRange searchRange) {
-    // searchStartOffset is used to find the closest match. caretOffset is used to reset the caret if there is no match.
-    // If searching based on e.g. :%s/... then these values are not going to be the same
+  public int updateIncsearchHighlights(@NotNull Editor editor, @NotNull String pattern, boolean forwards, int caretOffset, @Nullable LineRange searchRange) {
     final int searchStartOffset = searchRange != null ? EditorHelper.getLineStartOffset(editor, searchRange.getStartLine()) : caretOffset;
     final boolean showHighlights = OptionsManager.INSTANCE.getHlsearch().isSet();
-    int currentMatchOffset = updateSearchHighlights(pattern, false, showHighlights, searchStartOffset, searchRange, forwards, false);
-    LogicalPosition scrollTo = editor.offsetToLogicalPosition(currentMatchOffset == -1 ? caretOffset : currentMatchOffset);
-    editor.getScrollingModel().scrollTo(scrollTo, ScrollType.MAKE_VISIBLE);
+    return updateSearchHighlights(pattern, false, showHighlights, searchStartOffset, searchRange, forwards, false);
   }
 
   private void updateSearchHighlights(@Nullable String pattern, boolean shouldIgnoreSmartCase, boolean showHighlights, boolean forceUpdate) {
@@ -385,7 +399,7 @@ public class SearchGroup {
           removeSearchHighlight(editor);
         }
 
-        if (shouldAddSearchHighlight(editor, pattern, showHighlights)) {
+        if (shouldAddAllSearchHighlights(editor, pattern, showHighlights)) {
           final int startLine = searchRange == null ? 0 : searchRange.getStartLine();
           final int endLine = searchRange == null ? -1 : searchRange.getEndLine();
           List<TextRange> results = findAll(editor, pattern, startLine, endLine, shouldIgnoreCase(pattern, shouldIgnoreSmartCase));
@@ -395,15 +409,23 @@ public class SearchGroup {
           }
           UserDataManager.setVimLastSearch(editor, pattern);
         }
-        else if (!showHighlights && initialOffset != -1) {
-          // Incremental search always highlights current match. We know it's incsearch if we have a valid initial offset
+        else if (shouldAddCurrentMatchSearchHighlight(pattern, showHighlights, initialOffset)) {
           final boolean wrap = OptionsManager.INSTANCE.getWrapscan().isSet();
-          final TextRange result = findIt(editor, pattern, initialOffset, 1,
-            forwards ? DIR_FORWARDS : DIR_BACKWARDS, shouldIgnoreSmartCase, wrap, false, true);
+          final EnumSet<SearchOptions> searchOptions = EnumSet.of(SearchOptions.WHOLE_FILE);
+          if (wrap) searchOptions.add(SearchOptions.WRAP);
+          if (shouldIgnoreSmartCase) searchOptions.add(SearchOptions.IGNORE_SMARTCASE);
+          if (!forwards) searchOptions.add(SearchOptions.BACKWARDS);
+          final TextRange result = findIt(editor, pattern, initialOffset, 1, searchOptions);
           if (result != null && pattern != null) {
             currentMatchOffset = result.getStartOffset();
             final List<TextRange> results = Collections.singletonList(result);
             highlightSearchResults(editor, pattern, results, currentMatchOffset);
+          }
+        }
+        else if (shouldMaintainCurrentMatchOffset(pattern, initialOffset)) {
+          final Integer offset = UserDataManager.getVimIncsearchCurrentMatchOffset(editor);
+          if (offset != null) {
+            currentMatchOffset = offset;
           }
         }
       }
@@ -424,8 +446,33 @@ public class SearchGroup {
    * Add search highlights if hlSearch is true and the pattern is changed
    */
   @Contract("_, _, false -> false; _, null, true -> false")
-  private boolean shouldAddSearchHighlight(@NotNull Editor editor, @Nullable String newPattern, boolean hlSearch) {
+  private boolean shouldAddAllSearchHighlights(@NotNull Editor editor, @Nullable String newPattern, boolean hlSearch) {
     return hlSearch && newPattern != null && !newPattern.equals(UserDataManager.getVimLastSearch(editor)) && !Objects.equals(newPattern, "");
+  }
+
+  /**
+   * Add search highlight for current match if hlsearch is false and we're performing incsearch highlights
+   */
+  @Contract("_, true, _ -> false")
+  private boolean shouldAddCurrentMatchSearchHighlight(@Nullable String pattern, boolean hlSearch, int initialOffset) {
+    return !hlSearch && isIncrementalSearchHighlights(initialOffset) && pattern != null && pattern.length() > 0;
+  }
+
+  /**
+   * Keep the current match offset if the pattern is still valid and we're performing incremental search highlights
+   * This will keep the caret position when editing the offset in e.g. `/foo/e+1`
+   */
+  @Contract("null, _ -> false")
+  private boolean shouldMaintainCurrentMatchOffset(@Nullable String pattern, int initialOffset) {
+    return pattern != null && pattern.length() > 0 && isIncrementalSearchHighlights(initialOffset);
+  }
+
+  /**
+   * initialOffset is only valid if we're highlighting incsearch
+   */
+  @Contract(pure = true)
+  private boolean isIncrementalSearchHighlights(int initialOffset) {
+    return initialOffset != -1;
   }
 
   private void highlightSearchLines(@NotNull Editor editor, int startLine, int endLine) {
@@ -490,7 +537,8 @@ public class SearchGroup {
   @Nullable
   private TextRange findNextSearchForGn(@NotNull Editor editor, int count, boolean forwards) {
     if (forwards) {
-      return findIt(editor, lastSearch, editor.getCaretModel().getOffset(), count, DIR_FORWARDS, false, true, false, true);
+      final EnumSet<SearchOptions> searchOptions = EnumSet.of(SearchOptions.WRAP, SearchOptions.WHOLE_FILE);
+      return findIt(editor, lastSearch, editor.getCaretModel().getOffset(), count, searchOptions);
     } else {
       return searchBackward(editor, editor.getCaretModel().getOffset(), count);
     }
@@ -506,11 +554,13 @@ public class SearchGroup {
   @Nullable
   private TextRange searchBackward(@NotNull Editor editor, int offset, int count) {
     // Backward search returns wrongs end offset for some cases. That's why we should perform additional forward search
-    final TextRange foundBackward = findIt(editor, lastSearch, offset, count, DIR_BACKWARDS, false, true, false, true);
+    final EnumSet<SearchOptions> searchOptions = EnumSet.of(SearchOptions.WRAP, SearchOptions.WHOLE_FILE, SearchOptions.BACKWARDS);
+    final TextRange foundBackward = findIt(editor, lastSearch, offset, count, searchOptions);
     if (foundBackward == null) return null;
     int startOffset = foundBackward.getStartOffset() - 1;
     if (startOffset < 0) startOffset = EditorHelper.getFileSize(editor);
-    return findIt(editor, lastSearch, startOffset, 1, DIR_FORWARDS, false, true, false, true);
+    searchOptions.remove(SearchOptions.BACKWARDS);
+    return findIt(editor, lastSearch, startOffset, 1, searchOptions);
   }
 
   private static int distance(@NotNull TextRange range, int pos, boolean forwards, int size) {
@@ -523,19 +573,18 @@ public class SearchGroup {
     }
   }
 
-  @SuppressWarnings("SameParameterValue")
-  @Nullable
-  private static TextRange findIt(@NotNull Editor editor, @Nullable String pattern, int startOffset, int count, int dir,
-                                 boolean ignoreSmartCase, boolean wrap, boolean showMessages, boolean wholeFile) {
+  private static TextRange findIt(@NotNull Editor editor, @Nullable String pattern, int startOffset, int count, EnumSet<SearchOptions> searchOptions) {
     if (pattern == null || pattern.length() == 0) {
       logger.warn("Pattern is null or empty. Cannot perform search");
       return null;
     }
 
+    int dir = searchOptions.contains(SearchOptions.BACKWARDS) ? DIR_BACKWARDS : DIR_FORWARDS;
+
     //RE sp;
     RegExp sp;
     RegExp.regmmatch_T regmatch = new RegExp.regmmatch_T();
-    regmatch.rmm_ic = shouldIgnoreCase(pattern, ignoreSmartCase);
+    regmatch.rmm_ic = shouldIgnoreCase(pattern, searchOptions.contains(SearchOptions.IGNORE_SMARTCASE));
     sp = new RegExp();
     regmatch.regprog = sp.vim_regcomp(pattern, 1);
     if (regmatch.regprog == null) {
@@ -563,16 +612,16 @@ public class SearchGroup {
     //RegExp.regmmatch_T regmatch;
     CharPointer ptr;
     int matchcol;
-    int startcol;
+    RegExp.lpos_T matchpos;
     RegExp.lpos_T endpos = new RegExp.lpos_T();
     int loop;
     RegExp.lpos_T start_pos;
     boolean at_first_line;
-    int extra_col = 1;
+    int extra_col = dir == DIR_FORWARDS ? 1 : 0;
     boolean match_ok;
     long nmatched;
     //int         submatch = 0;
-    int first_lnum;
+    boolean first_match = true;
 
     int lineCount = EditorHelper.getLineCount(editor);
     int startLine = 0;
@@ -602,7 +651,7 @@ public class SearchGroup {
 
       int lcount = EditorHelper.getLineCount(editor);
       for (loop = 0; loop <= 1; ++loop)   /* loop twice if 'wrapscan' set */ {
-        if (!wholeFile) {
+        if (!searchOptions.contains(SearchOptions.WHOLE_FILE)) {
           startLine = lnum;
           endLine = lnum + 1;
         }
@@ -610,14 +659,13 @@ public class SearchGroup {
           /*
           * Look for a match somewhere in the line.
           */
-          first_lnum = lnum;
           nmatched = sp.vim_regexec_multi(regmatch, editor, lcount, lnum, 0);
           if (nmatched > 0) {
             /* match may actually be in another line when using \zs */
-            lnum += regmatch.startpos[0].lnum;
-            ptr = new CharPointer(EditorHelper.getLineBuffer(editor, lnum));
-            startcol = regmatch.startpos[0].col;
+            matchpos = new RegExp.lpos_T(regmatch.startpos[0]);
             endpos = new RegExp.lpos_T(regmatch.endpos[0]);
+
+            ptr = new CharPointer(EditorHelper.getLineBuffer(editor, lnum + matchpos.lnum));
 
             /*
             * Forward search in the first line: match should be after
@@ -631,16 +679,19 @@ public class SearchGroup {
               * one back afterwards, compare with that position,
               * otherwise "/$" will get stuck on end of line.
               */
-              while ((startcol - (startcol == ptr.strlen() ? 1 : 0)) < (start_pos.col + extra_col)) {
+              while (matchpos.lnum == 0
+                && (searchOptions.contains(SearchOptions.WANT_ENDPOS) && first_match
+                  ? (nmatched == 1 && endpos.col - 1 < start_pos.col + extra_col)
+                  : (matchpos.col - (ptr.charAt(matchpos.col) == '\u0000' ? 1 : 0) < start_pos.col + extra_col))) {
                 if (nmatched > 1) {
                   /* end is in next line, thus no match in
-               * this line */
+                   * this line */
                   match_ok = false;
                   break;
                 }
                 matchcol = endpos.col;
                 /* for empty match: advance one char */
-                if (matchcol == startcol && ptr.charAt(matchcol) != '\u0000') {
+                if (matchcol == matchpos.col && ptr.charAt(matchcol) != '\u0000') {
                   ++matchcol;
                 }
                 if (ptr.charAt(matchcol) == '\u0000' ||
@@ -648,11 +699,11 @@ public class SearchGroup {
                   match_ok = false;
                   break;
                 }
-                startcol = regmatch.startpos[0].col;
+                matchpos = new RegExp.lpos_T(regmatch.startpos[0]);
                 endpos = new RegExp.lpos_T(regmatch.endpos[0]);
 
                 /* Need to get the line pointer again, a
-        * multi-line search may have made it invalid. */
+                 * multi-line search may have made it invalid. */
                 ptr = new CharPointer(EditorHelper.getLineBuffer(editor, lnum));
               }
               if (!match_ok) {
@@ -668,12 +719,15 @@ public class SearchGroup {
               * relative to the end of the match.
               */
               match_ok = false;
-              for (; ; ) {
-                if (!at_first_line || (CharacterPosition.charOffsetOnLineToColumn(editor, lnum, regmatch.startpos[0].col + extra_col) <= start_pos.col)) {
+              for (;;) {
+                if (loop != 0 ||
+                  (searchOptions.contains(SearchOptions.WANT_ENDPOS)
+                    ? (lnum + regmatch.endpos[0].lnum < start_pos.lnum || (lnum + regmatch.endpos[0].lnum == start_pos.lnum && regmatch.endpos[0].col - 1 < start_pos.col + extra_col))
+                    : (lnum + regmatch.startpos[0].lnum < start_pos.lnum || (lnum + regmatch.startpos[0].lnum == start_pos.lnum && regmatch.startpos[0].col < start_pos.col + extra_col)))) {
                   /* Remember this position, we use it if it's
-           * the last match in the line. */
+                   * the last match in the line. */
                   match_ok = true;
-                  startcol = regmatch.startpos[0].col;
+                  matchpos = new RegExp.lpos_T(regmatch.startpos[0]);
                   endpos = new RegExp.lpos_T(regmatch.endpos[0]);
                 }
                 else {
@@ -692,17 +746,17 @@ public class SearchGroup {
                 }
                 matchcol = endpos.col;
                 /* for empty match: advance one char */
-                if (matchcol == startcol && ptr.charAt(matchcol) != '\u0000') {
+                if (matchcol == matchpos.col && ptr.charAt(matchcol) != '\u0000') {
                   ++matchcol;
                 }
                 if (ptr.charAt(matchcol) == '\u0000' ||
-                    (nmatched = sp.vim_regexec_multi(regmatch, editor, lcount, lnum, matchcol)) == 0) {
+                    (nmatched = sp.vim_regexec_multi(regmatch, editor, lcount, lnum + matchpos.lnum, matchcol)) == 0) {
                   break;
                 }
 
                 /* Need to get the line pointer again, a
-        * multi-line search may have made it invalid. */
-                ptr = new CharPointer(EditorHelper.getLineBuffer(editor, lnum));
+                 * multi-line search may have made it invalid. */
+                ptr = new CharPointer(EditorHelper.getLineBuffer(editor, lnum + matchpos.lnum));
               }
 
               /*
@@ -714,10 +768,11 @@ public class SearchGroup {
               }
             }
 
-            pos.lnum = lnum;
-            pos.col = startcol;
-            endpos.lnum += first_lnum;
+            pos.lnum = lnum + matchpos.lnum;
+            pos.col = matchpos.col;
+            endpos.lnum = lnum + endpos.lnum;
             found = 1;
+            first_match = false;
 
             /* Set variables used for 'incsearch' highlighting. */
             //search_match_lines = endpos.lnum - (lnum - first_lnum);
@@ -738,7 +793,7 @@ public class SearchGroup {
         * stop the search if wrapscan isn't set, after an interrupt and
         * after a match
         */
-        if (!wrap || found != 0) {
+        if (!searchOptions.contains(SearchOptions.WRAP) || found != 0) {
           break;
         }
 
@@ -767,8 +822,8 @@ public class SearchGroup {
 
     if (found == 0)             /* did not find it */ {
       //if ((options & SEARCH_MSG) == SEARCH_MSG)
-      if (showMessages) {
-        if (wrap) {
+      if (searchOptions.contains(SearchOptions.SHOW_MESSAGES)) {
+        if (searchOptions.contains(SearchOptions.WRAP)) {
           VimPlugin.showMessage(MessageHelper.message(Msg.e_patnotf2, pattern));
         }
         else if (lnum <= 0) {
@@ -802,73 +857,94 @@ public class SearchGroup {
       final RangeHighlighter highlighter = highlightMatch(editor, range.getStartOffset(), range.getEndOffset(), current, pattern);
       highlighters.add(highlighter);
     }
+
+    UserDataManager.setVimIncsearchCurrentMatchOffset(editor, currentMatchOffset);
   }
 
   private int findItOffset(@NotNull Editor editor, int startOffset, int count, int dir) {
     boolean wrap = OptionsManager.INSTANCE.getWrapscan().isSet();
     logger.info("Perform search. Direction: " + dir + " wrap: " + wrap);
-    TextRange range = findIt(editor, lastSearch, startOffset, count, dir, lastIgnoreSmartCase, wrap, true, true);
-    if (range == null) {
-      logger.warn("No range is found");
-      return -1;
-    }
+
+    int offset = 0;
+    boolean offsetIsLineOffset = false;
+    boolean hasEndOffset = false;
 
     ParsePosition pp = new ParsePosition(0);
-    int res = range.getStartOffset();
 
     if (lastOffset == null) {
       logger.warn("Last offset is null. Cannot perform search");
       return -1;
     }
 
-    if (lastOffset.length() == 0) {
-      return range.getStartOffset();
-    }
-    else if (Character.isDigit(lastOffset.charAt(0)) || lastOffset.charAt(0) == '+' || lastOffset.charAt(0) == '-') {
-      int lineOffset = 0;
-      if (lastOffset.equals("+")) {
-        lineOffset = 1;
-      }
-      else if (lastOffset.equals("-")) {
-        lineOffset = -1;
-      }
-      else {
-        if (lastOffset.charAt(0) == '+') {
-          lastOffset = lastOffset.substring(1);
-        }
-        NumberFormat nf = NumberFormat.getIntegerInstance();
-        pp = new ParsePosition(0);
-        Number num = nf.parse(lastOffset, pp);
-        if (num != null) {
-          lineOffset = num.intValue();
-        }
-      }
+    if (lastOffset.length() > 0) {
+      if (Character.isDigit(lastOffset.charAt(0)) || lastOffset.charAt(0) == '+' || lastOffset.charAt(0) == '-') {
+        offsetIsLineOffset = true;
 
+        if (lastOffset.equals("+")) {
+          offset = 1;
+        } else if (lastOffset.equals("-")) {
+          offset = -1;
+        } else {
+          if (lastOffset.charAt(0) == '+') {
+            lastOffset = lastOffset.substring(1);
+          }
+          NumberFormat nf = NumberFormat.getIntegerInstance();
+          pp = new ParsePosition(0);
+          Number num = nf.parse(lastOffset, pp);
+          if (num != null) {
+            offset = num.intValue();
+          }
+        }
+      } else if ("ebs".indexOf(lastOffset.charAt(0)) != -1) {
+        if (lastOffset.length() >= 2) {
+          if ("+-".indexOf(lastOffset.charAt(1)) != -1) {
+            offset = 1;
+          }
+          NumberFormat nf = NumberFormat.getIntegerInstance();
+          pp = new ParsePosition(lastOffset.charAt(1) == '+' ? 2 : 1);
+          Number num = nf.parse(lastOffset, pp);
+          if (num != null) {
+            offset = num.intValue();
+          }
+        }
+
+        hasEndOffset = lastOffset.charAt(0) == 'e';
+      }
+    }
+
+    /*
+     * If there is a character offset, subtract it from the current
+     * position, so we don't get stuck at "?pat?e+2" or "/pat/s-2".
+     * Skip this if pos.col is near MAXCOL (closed fold).
+     * This is not done for a line offset, because then we would not be vi
+     * compatible.
+     */
+    if (!offsetIsLineOffset && offset != 0) {
+      startOffset = Math.max(0, Math.min(startOffset - offset, EditorHelper.getFileSize(editor) - 1));
+    }
+
+    EnumSet<SearchOptions> searchOptions = EnumSet.of(SearchOptions.SHOW_MESSAGES, SearchOptions.WHOLE_FILE);
+    if (dir == DIR_BACKWARDS) searchOptions.add(SearchOptions.BACKWARDS);
+    if (lastIgnoreSmartCase) searchOptions.add(SearchOptions.IGNORE_SMARTCASE);
+    if (wrap) searchOptions.add(SearchOptions.WRAP);
+    if (hasEndOffset) searchOptions.add(SearchOptions.WANT_ENDPOS);
+    TextRange range = findIt(editor, lastSearch, startOffset, count, searchOptions);
+    if (range == null) {
+      logger.warn("No range is found");
+      return -1;
+    }
+
+    int res = range.getStartOffset();
+
+    if (offsetIsLineOffset) {
       int line = editor.offsetToLogicalPosition(range.getStartOffset()).line;
-      int newLine = EditorHelper.normalizeLine(editor, line + lineOffset);
+      int newLine = EditorHelper.normalizeLine(editor, line + offset);
 
       res = VimPlugin.getMotion().moveCaretToLineStart(editor, newLine);
     }
-    else if ("ebs".indexOf(lastOffset.charAt(0)) != -1) {
-      int charOffset = 0;
-      if (lastOffset.length() >= 2) {
-        if ("+-".indexOf(lastOffset.charAt(1)) != -1) {
-          charOffset = 1;
-        }
-        NumberFormat nf = NumberFormat.getIntegerInstance();
-        pp = new ParsePosition(lastOffset.charAt(1) == '+' ? 2 : 1);
-        Number num = nf.parse(lastOffset, pp);
-        if (num != null) {
-          charOffset = num.intValue();
-        }
-      }
-
-      int base = range.getStartOffset();
-      if (lastOffset.charAt(0) == 'e') {
-        base = range.getEndOffset() - 1;
-      }
-
-      res = Math.max(0, Math.min(base + charOffset, EditorHelper.getFileSize(editor) - 1));
+    else if (hasEndOffset || offset != 0) {
+      int base = hasEndOffset ? range.getEndOffset() - 1 : range.getStartOffset();
+      res = Math.max(0, Math.min(base + offset, EditorHelper.getFileSize(editor) - 1));
     }
 
     int ppos = pp.getIndex();
@@ -1165,7 +1241,7 @@ public class SearchGroup {
           if (do_ask) {
             RangeHighlighter hl = highlightConfirm(editor, startoff, endoff);
             MotionGroup.scrollPositionIntoView(editor, editor.offsetToVisualPosition(startoff), true);
-            MotionGroup.moveCaret(editor, caret, start);
+            MotionGroup.moveCaret(editor, caret, startoff);
             final ReplaceConfirmationChoice choice = confirmChoice(editor, match);
             editor.getMarkupModel().removeHighlighter(hl);
             switch (choice) {
@@ -1224,9 +1300,10 @@ public class SearchGroup {
     }
 
     if (lastMatch != -1) {
-      MotionGroup.moveCaret(editor, caret, VimPlugin.getMotion().moveCaretToLineStartSkipLeading(editor,
-                                                                                                 editor.offsetToLogicalPosition(
-                                                                                                     lastMatch).line));
+      if (!got_quit) {
+        MotionGroup.moveCaret(editor, caret,
+          VimPlugin.getMotion().moveCaretToLineStartSkipLeading(editor, editor.offsetToLogicalPosition(lastMatch).line));
+      }
     }
     else {
       VimPlugin.showMessage(MessageHelper.message(Msg.e_patnotf2, pattern));
@@ -1256,13 +1333,32 @@ public class SearchGroup {
       attributes.setEffectType(EffectType.ROUNDED_BOX);
       attributes.setEffectColor(editor.getColorsScheme().getColor(EditorColors.CARET_COLOR));
     }
+    if (attributes.getErrorStripeColor() == null) {
+      attributes.setErrorStripeColor(getFallbackErrorStripeColor(attributes, editor.getColorsScheme()));
+    }
     final RangeHighlighter highlighter = editor.getMarkupModel().addRangeHighlighter(start, end, HighlighterLayer.SELECTION - 1,
       attributes, HighlighterTargetArea.EXACT_RANGE);
     highlighter.setErrorStripeTooltip(tooltip);
     return highlighter;
   }
 
+  /**
+   * Return a valid error stripe colour based on editor background
+   *
+   * Based on HighlightManager#addRangeHighlight behaviour, which we can't use because it will hide highlights when
+   * hitting Escape
+   */
+  private static @Nullable Color getFallbackErrorStripeColor(TextAttributes attributes, EditorColorsScheme colorsScheme) {
+    if (attributes.getBackgroundColor() != null) {
+      boolean isDark = ColorUtil.isDark(colorsScheme.getDefaultBackground());
+      return isDark ? attributes.getBackgroundColor().brighter() : attributes.getBackgroundColor().darker();
+    }
+    return null;
+  }
+
   private static void removeSearchHighlight(@NotNull Editor editor) {
+    UserDataManager.setVimLastSearch(editor, null);
+
     Collection<RangeHighlighter> ehl = UserDataManager.getVimLastHighlighters(editor);
     if (ehl == null) {
       return;
@@ -1275,7 +1371,6 @@ public class SearchGroup {
     ehl.clear();
 
     UserDataManager.setVimLastHighlighters(editor, null);
-    UserDataManager.setVimLastSearch(editor, null);
   }
 
   public void saveData(@NotNull Element element) {
@@ -1348,6 +1443,7 @@ public class SearchGroup {
   /**
    * Updates search highlights when the selected editor changes
    */
+  @SuppressWarnings("unused")
   public static void fileEditorManagerSelectionChangedCallback(@NotNull FileEditorManagerEvent event) {
     VimPlugin.getSearch().updateSearchHighlights();
   }
@@ -1409,6 +1505,15 @@ public class SearchGroup {
     SKIP,
     QUIT,
     SUBSTITUTE_ALL,
+  }
+
+  private enum SearchOptions {
+    BACKWARDS,
+    WANT_ENDPOS,
+    WRAP,
+    SHOW_MESSAGES,
+    WHOLE_FILE,
+    IGNORE_SMARTCASE,
   }
 
   @Nullable private String lastSearch;